% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anothertrial.R
\name{anothertrial}
\alias{anothertrial}
\title{Second attempt at metatrial function}
\usage{
anothertrial(true_median = 50, ...)
}
\arguments{
\item{true_effect}{The value of the control population median.}
}
\description{
This time, let's try making function so that it can be wrapped in safely at
the metasim level.
}
