% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_par.R
\name{beta_par}
\alias{beta_par}
\title{calculate beta parameters}
\usage{
beta_par(proportion, error)
}
\arguments{
\item{proportion}{Expected proportion for the intervention group}

\item{error}{Within what value will 90 per cent of the proportion of intervention groups fall within?}
}
\description{
calculate beta parameters
}
