% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_sim_pars.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\title{Default Simulation Parameters}
\format{A dataframe with 6 rows and 2 variables.
\describe{
\item{dist}{R-friendly distribution name.}
\item{par}{Parameter set for r\code{dist} random sampling.}
}}
\usage{
default_parameters
}
\description{
These are the default distributional parameter sets used in \link{metasim}.
}
\keyword{datasets}
