% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_name.R
\name{dist_name}
\alias{dist_name}
\title{R to name}
\usage{
dist_name(dist, toR = FALSE)
}
\arguments{
\item{dist}{A string with a distribution name, e.g. "lnorm" or "log-normal".}

\item{toR}{Boolean as to whether it's to the R name or not. Defaults to
false.}
}
\description{
A function that switches out the R name for a distribution with the real
name, and back again.
}
