% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_se.R
\name{effect_se}
\alias{effect_se}
\title{estimator for variance of sample median}
\usage{
effect_se(centre, spread, centre_type = "median", spread_type = "iqr",
  n = 1)
}
\arguments{
\item{centre}{a mean or median}

\item{spread}{a sd, var, iqr, or range}

\item{centre_type}{specify "mean" or "median"}

\item{spread_type}{specify "sd", "var", "iqr", or "range"}
}
\description{
estimator for variance of sample median
}
