% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_comparison.R
\name{estimator_comparison}
\alias{estimator_comparison}
\title{Plot ratio-comparisons}
\usage{
estimator_comparison()
}
\description{
Example plot, by rights should be in varameta - todo:
}
\details{
Right now it' convenient to have this function here.
}
