% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lnorm.R
\name{g_lnorm}
\alias{g_lnorm}
\title{Estimator, assuming a log-normal distribution.}
\usage{
g_lnorm(median, spread, n = 1, spread_type = "iqr")
}
\arguments{
\item{median}{sample median}

\item{spread}{iqr or range value}

\item{n}{sample size}

\item{spread_type}{iqr or range, defaults to iqr}
}
\description{
This estimator is one choice for density substitution in the variance of the sample median estimator.
}
