% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parameters.R
\name{get_arguments}
\alias{get_arguments}
\title{Simulation-level parameters}
\usage{
get_arguments(...)
}
\description{
Simulation-level parameters
}
\examples{
get_arguments()
get_arguments(a = 3, x = "1")

}
