% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parameters.R
\name{get_defaults}
\alias{get_defaults}
\title{Get simulation-level parameters}
\usage{
get_defaults(...)
}
\description{
Get default arguments for simulation
}
\examples{
# check defaults
get_defaults()
# does it still return defaults?
get_defaults(a = 3)
}
