% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parameters.R
\name{get_specified}
\alias{get_specified}
\title{Get specified argments}
\usage{
get_specified(...)
}
\description{
Get specified argments
}
\examples{
get_specified()
get_specified(a = 3)
get_specified(a = 3, b = "hello, world")
}
