% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_proportion.R
\name{intervention_proportion}
\alias{intervention_proportion}
\title{Calculate proportion of intervention group}
\usage{
intervention_proportion(n, proportion, error)
}
\arguments{
\item{proportion}{Expected proportion for the intervention group}

\item{error}{Within what value will 90 per cent of the proportion of intervention groups fall within?}
}
\description{
Calculate proportion of intervention group
}
