% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_se.R
\name{lr_se}
\alias{lr_se}
\title{calculate the standard error of the log-ratio of two random samples}
\usage{
lr_se(measure, n_c, effect_c, effect_se_c, n_i, effect_i, effect_se_i)
}
\arguments{
\item{measure}{mean or median}

\item{n_c}{sample size of control group}

\item{effect_c}{effect measure for control group}

\item{effect_se_c}{standard error of control group}

\item{effect_i}{effect measure of intervention group}

\item{effect_se_i}{standard error of intervention group}

\item{n_c}{sample size of intervention group}
}
\description{
Provides sample estimator for the squared-root of the variance of the log-ratio of two random samples
}
