% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafor_converge.R
\docType{data}
\name{metafor_converge}
\alias{metafor_converge}
\title{metafor convergence examples}
\format{A dataframe.
\describe{
\item{yi}{effect}
\item{vi}{variance of the effect}
}}
\source{
{www.metafor-project.org/doku.php/tips:convergence_problems_rmaa}
}
\usage{
metafor_converge
}
\description{
Effects and variance of the effect from the examples provided in \code{\link{metafor}::} documentation of convergence problems with \code{\link{rma}} function. The convergent data set corresponds to Example 1, long_converge to Example 2, and adj_step to Example 3.
}
\keyword{datasets}
