% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamodel.R
\name{metamodel}
\alias{metamodel}
\title{Meta-analyse with error handling}
\usage{
metamodel(y = rnorm(10, mean = 20, sd = 0.2), se = runif(10, 0, 0.8),
  test = "knha", data = NULL)
}
\arguments{
\item{test}{"knha" or "z"}
}
\description{
If random-effects model fails to converge, then replace with fixed-effects model,
else, produce error string.
}
