% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasim.R
\name{metasim}
\alias{metasim}
\title{one row, one simulations}
\usage{
metasim(..., id = "simulation1", trial_fn = metatrial, trials = 4)
}
\arguments{
\item{...}{\code{trial_fn} arguments, i.e., simulation nparameters}

\item{trial_fn}{the function to repeat}

\item{trials}{the number of trials per simulation}
}
\description{
runs on one row, returns coverage probability
}
