% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlesim_plot.R
\name{metasim_plot}
\alias{metasim_plot}
\title{plot coverage probability}
\usage{
metasim_plot(sim = meta_sim)
}
\arguments{
\item{sim}{The output of the \link{metasims} function.}
}
\description{
Plot single-study simulation data
}
\details{
Expects input of the return of \link{metasims} (with default \code{single_study = FALSE}).
}
