% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastats.R
\name{metastats}
\alias{metastats}
\title{simulate a meta-analysis dataset}
\usage{
metastats(n_df, sampling_fn = rnorm, fn_parameters = list(mean = 50, sd
  = 0.2), between_study_variation = 0.4, within_study_variation = 0.2,
  median_ratio = 1.2)
}
\arguments{
\item{n_df}{\code{data.frame} of sample sizes, such as output by \code{\link{meta_n}}.}

\item{sample_fn}{}
}
\description{
simulate a meta-analysis dataset
}
