% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-methods.R
\name{plot.sim_ss}
\alias{plot.sim_ss}
\title{Plot coverage probability for single-study simulation}
\usage{
\method{plot}{sim_ss}(sim = single_sim)
}
\arguments{
\item{sim}{A \code{sim_ss} object from \link{metasim}.}
}
\description{
Plot single-study simulation results.
}
