% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-methods.R
\name{plots}
\alias{plots}
\alias{sim_parameter}
\alias{caption_text}
\alias{caption.sim_ma}
\alias{caption.sim_ss}
\alias{plot.sim_ss}
\title{A first pass at S3 methods for simulation objects}
\usage{
sim_parameter(sim, arg)

caption_text(sim)

\method{caption}{sim_ma}(sim)

\method{caption}{sim_ss}(sim)

\method{plot}{sim_ss}(sim = single_sim, caption = NULL)
}
\arguments{
\item{sim}{A \code{sim_ss} object from \link{metasim}.}

\item{caption}{Whether or not to include the caption.}
}
\description{
Methods for simulation results.

Plot single-study simulation results.
}
