% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_plot.R
\name{rho_plot}
\alias{rho_plot}
\title{plot the distribution of the proportion allocated to the intervention group}
\usage{
rho_plot(mu, epsilon)
}
\description{
plot the distribution of the proportion allocated to the intervention group
}
