% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df}
\alias{sim_df}
\title{generate simulation parameter dataframe}
\usage{
sim_df()
}
\arguments{
\item{dist_tribble}{A \code{\link{tibble::tribble}} with one column for distribution, and one column for the parameters.}

\item{k}{Simulate for different numbers of studies.}

\item{between_study_variation}{tau}

\item{within_study_variation}{epsilon}

\item{prop}{proportion of sample size we expect to cohorts to vary by most of the time}
}
\description{
generate simulation parameter dataframe
}
