% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlesim_plot.R
\name{singlesim_caption}
\alias{singlesim_caption}
\title{caption distributions}
\usage{
singlesim_caption(sim, distributions = default_parameters,
  cap_width = 110, measure = "median", metasims_args = args(metasims)
  \%>\% as.list())
}
\arguments{
\item{sim}{The output of the \link{metasims} function.}
}
\description{
caption for \link{metasim_plot}.
}
