% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singletrial.R
\name{singletrial}
\alias{singletrial}
\title{Single-study trial for coverage of medians}
\usage{
singletrial(measure = "median", measure_spread = "iqr", tau_sq = 0,
  effect_ratio = 1.2, rdist = "norm", parameters = list(mean = 50, sd
  = 0.2), n_df = sim_n(k = 1), true_effect = 50, knha = TRUE)
}
\description{
This trial samples one paired set of data and returns the confidence interval, bias, and coverage.
}
\details{
NB: bias is effect - true effect.
}
