% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singletrial_test.R
\name{singletrial_test}
\alias{singletrial_test}
\title{testing singletrial function}
\usage{
singletrial_test(tau_sq = 0, median_ratio = 1.2, rdist = "norm",
  parameters = list(mean = 50, sd = 0.2), n_df = sim_n(k = 1),
  true_effect = 50, knha = TRUE)
}
\description{
testing singletrial function
}
