% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sim.R
\name{tidy_sim}
\alias{tidy_sim}
\title{useful simulation statistics}
\usage{
tidy_sim(rma_model)
}
\description{
Returns statistics from what are currently \link[metabroom:glance]{metabroom::glance}
and \link[metabroom:tidy]{metabroom::tidy} functions that are used in the \link[metasim:metasim]{metasim::metasim}
package for analysis of simulation results.
}
\details{
todo: this will port to broom:: proper eventually.
}
