% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toss.R
\name{toss}
\alias{toss}
\title{Toss away bad simulated data}
\usage{
toss(code, error_msg = NULL, warning_msg = NULL)
}
\description{
This function returns a function's output if the function runs, or allows the
user to customise the error and warning messages if the function throws
either.
}
