% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta_plot.R
\name{zeta_plot}
\alias{zeta_plot}
\title{plot the distribution of the proportion allocated to the intervention group}
\usage{
zeta_plot(mu, epsilon)
}
\description{
plot the distribution of the proportion allocated to the intervention group
}
