% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anothertrial.R
\name{anothertrial}
\alias{anothertrial}
\title{Second attempt at metatrial function}
\usage{
anothertrial(true_median = 50, tau_sq = 0.6, median_ratio = 1.2,
  rdist = "norm", parameters = list(mean = 50, sd = 0.2),
  n_df = sim_n(k = 3), knha = TRUE, true_effect = 50,
  test = "knha")
}
\arguments{
\item{median_ratio}{ratio of population medians}

\item{rdist}{string indicating distribution, "norm", "lnorm", "exp", or "pareto"}

\item{n_df}{\code{data.frame} of sample sizes,
such as output by \code{\link{meta_n}}.}

\item{true_effect}{The value of the control population median.}

\item{test}{"knha" or "z" for \link[metafor:rma]{metafor::rma}.}
}
\description{
This time, let's try making function so that it can be wrapped in safely at
the metasim level.
}
