% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasims.R
\name{metasims}
\alias{metasims}
\title{simulate over all}
\usage{
metasims(dist_tribble = tibble::tribble(~dist, ~par, "norm", list(mean =
  67, sd = 0.3), "exp", list(rate = 3), "pareto", list(shape = 3, scale =
  3), "pareto", list(shape = 2, scale = 1), "pareto", list(shape = 0.5,
  scale = 1), "lnorm", list(mean = 44, sd = 0.3)), k = c(3, 7, 50),
  between_study_variation = seq(0, 0.4, 0.2), median_ratio = c(1, 1.2),
  prop = 0.3, trials = 10, trial_fn = metatrial)
}
\arguments{
\item{k}{Simulate for different numbers of studies.}

\item{between_study_variation}{tau}

\item{prop}{proportion of sample size we expect to cohorts to vary by most of the time}

\item{within_study_variation}{epsilon}
}
\description{
simulate over all
}
