% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatrial.R
\name{metatrial}
\alias{metatrial}
\title{one trial}
\usage{
metatrial(tau = 0.6, median_ratio = 1.2, rdist = "norm",
  parameters = list(mean = 50, sd = 0.2), n_df = sim_n(k = 3),
  knha = TRUE, true_effect = 50)
}
\arguments{
\item{median_ratio}{ratio of population medians}

\item{rdist}{string indicating distribution, "norm", "lnorm", "exp", or "pareto"}

\item{n_df}{\code{data.frame} of sample sizes,
such as output by \code{\link{meta_n}}.}

\item{knha}{Knapp-Hartung test instead of the
default "z" test in `metafor::rma`.}
}
\description{
one trial
}
