% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatrial.R
\name{metatrial}
\alias{metatrial}
\title{one trial}
\usage{
metatrial(between_study_variation = 0.6, within_study_variation = 0.3,
  median_ratio = 1.2, rdist = "norm", parameters = list(mean = 50, sd
  = 0.2), n_df = sim_n(k = 3), knha = TRUE, true_effect = 50)
}
\arguments{
\item{knha}{Knapp-Hartung test instead of the default "z" test in `metafor::rma`.}
}
\description{
one trial
}
