% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatrial.R
\name{metatrial}
\alias{metatrial}
\title{Generate meta-analysis data and calculate estimator}
\usage{
metatrial(measure = "median", measure_spread = "iqr", tau_sq = 0.6,
  effect_ratio = 1.2, rdist = "norm", parameters = list(mean = 50, sd
  = 0.2), n_df = sim_n(k = 3), knha = TRUE, true_effect = 50,
  test = "knha")
}
\arguments{
\item{effect_ratio}{ratio of population effects intervention / control}

\item{rdist}{string indicating distribution, "norm", "lnorm", "exp", or "pareto"}

\item{n_df}{\code{data.frame} of sample sizes,
such as output by \code{\link{meta_n}}.}

\item{true_effect}{The value of the control population median.}

\item{test}{"knha" or "z" for \link[metafor:rma]{metafor::rma}.}
}
\description{
Simulate data based on simulation parameters and meta-analyse.
}
\details{
NB: bias is effect - true effect.
}
