% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-methods.R
\name{plots}
\alias{plots}
\alias{dat_parameter}
\alias{caption_text}
\alias{caption.sim_ma}
\alias{caption.sim_ss}
\alias{caption.distributions}
\alias{plot.sim_ss}
\alias{plot.sim_ma}
\alias{plot.dist}
\title{A first pass at S3 methods for datulation objects}
\usage{
dat_parameter(dat, arg)

caption_text(dat)

\method{caption}{sim_ma}(dat)

\method{caption}{sim_ss}(dat)

\method{caption}{distributions}(dat)

\method{plot}{sim_ss}(dat = single_sim, caption = NULL)

\method{plot}{sim_ma}(dat, caption = NULL)

\method{plot}{dist}(dat, caption = NULL)
}
\arguments{
\item{caption}{Whether or not to include the caption.}

\item{sim}{A \code{sim_ss} object from \link{metasim}.}
}
\description{
Methods for datulation results.

Plot single-study simulation results.
}
