% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{rma}
\alias{forest}
\alias{escalc}
\alias{pluck}
\alias{clean_names}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{janitor}{\code{\link[janitor]{clean_names}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{metafor}{\code{\link[metafor]{rma}}, \code{\link[metafor]{forest}}, \code{\link[metafor]{escalc}}}

  \item{purrr}{\code{\link[purrr]{pluck}}}
}}

