% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df}
\alias{sim_df}
\title{generate simulation parameter dataframe}
\usage{
sim_df(dist_tribble = default_parameters, k = c(3, 7, 20),
  tau2 = seq(0, 0.4, 0.2), median_ratio = c(1, 1.2), prop = 0.3)
}
\arguments{
\item{dist_tribble}{A \code{\link{tibble::tribble}}
with one column for distribution, and one column for the parameters}

\item{k}{number of studies, defaults to 3}

\item{median_ratio}{\eqn{\nu_I / \nu_C := \rho} where \eqn{\rho} denotes the ratio of medians.}

\item{prop}{proportion of n we expect}
}
\description{
generate simulation parameter dataframe
}
