% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df}
\alias{sim_df}
\title{generate simulation parameter dataframe}
\usage{
sim_df(dist_tribble = default_parameters, k = c(3, 7, 20),
  tau2 = seq(0, 0.4, by = 0.2), effect_ratio = c(1, 1.2), min_n = 20,
  max_n = 200, prop = 0.5, prop_error = 0.1)
}
\arguments{
\item{dist_tribble}{A \code{\link{tibble::tribble}}
with one column for distribution, and one column for the parameters}

\item{k}{number of studies, defaults to 3}

\item{min_n}{lower bound for sample sizes, defaults to 20}

\item{max_n}{upper bound for sample sizes, defaults to 200}

\item{prop}{proportion of n we expect}

\item{prop_error}{what do we expect 90% of proportions to fall within?}
}
\description{
generate simulation parameter dataframe
}
