% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sample.R
\name{sim_sample}
\alias{sim_sample}
\title{Simulate a sample}
\usage{
sim_sample(n, rdist, par)
}
\arguments{
\item{n}{sample size}

\item{rdist}{random sampling function, such as \code{\link{rnorm}} and \code{\link{rexp}}.}

\item{par}{list of parameter arguments}
}
\description{
Given a distribution, parameters, sample size, generate a sample.
}
