% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sample.R
\name{sim_sample}
\alias{sim_sample}
\title{Simulate a sample}
\usage{
sim_sample(n = 18, this_study_error = 0.2, rdist = "norm",
  par = list(mean = 20, sd = 0.2), control = TRUE,
  effect_ratio = 1.2)
}
\arguments{
\item{n}{sample size}

\item{this_study_error}{this study error}

\item{rdist}{string indicating distribution, "norm", "lnorm", "exp", or "pareto"}

\item{par}{list of parameter arguments}

\item{control}{value of first parameter of distribution is determined by median ratio}

\item{effect_ratio}{ratio of population effects intervention / control}

\item{epsilon}{within study error}
}
\description{
Given a distribution, parameters, sample size, generate a sample.
}
