% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_stats.R
\name{sim_stats}
\alias{sim_stats}
\title{Simulate a meta-analysis dataset}
\usage{
sim_stats(measure = "median", measure_spread = "iqr", n_df = sim_n(),
  wide = FALSE, rdist = "norm", par = list(mean = 50, sd = 0.2),
  tau_sq = 0.4, effect_ratio = 1.2)
}
\arguments{
\item{n_df}{\code{data.frame} of sample sizes,
such as output by \code{\link{meta_n}}.}

\item{wide}{Logical indicating if wide format, as is expected by \code{\link{metafor}::}.}

\item{rdist}{string indicating distribution, "norm", "lnorm", "exp", or "pareto"}

\item{par}{list of parameter arguments}

\item{effect_ratio}{ratio of population effects intervention / control}
}
\description{
Simulate a meta-analysis dataset
}
