% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toss.R
\name{toss}
\alias{toss}
\title{Toss away bad simulated data}
\usage{
toss(code, silent = FALSE)
}
\description{
This function (hopefully) will toss away crappy stuff from the
\code{\link{metatrial}} and metasim functions.
}
