% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toss.R
\name{toss}
\alias{toss}
\title{Toss away bad simulated data}
\usage{
toss(code, error_msg = NULL, warning_msg = NULL)
}
\description{
This function returns a function's output if the function runs, or allows the
user to customise the error and warning messages if the function throws
either.
}
\examples{

sum(c(1, 2, 3))

# an element of wrong type will cause sum to throw an error
\dontrun{
sum(c(1, 2, "a"))
}

# toss returns function output if function runs
toss(sum(c(1, 2, 3)))

# toss returns NULL by default if function does not run
toss(sum(c(1, 2, "a")))

# toss allows you to customise the error message
toss(sum(c(1, 2, "a")), error_msg = "the cake is a lie")

}
